move namespace package protection interface into cop
