+++ lib/package_protections/private.rb
@@ -7,7 +7,6 @@ require 'package_protections/private/typed_api_protection'
 require 'package_protections/private/incoming_privacy_protection'
 require 'package_protections/private/outgoing_dependency_protection'
 require 'package_protections/private/metadata_modifiers'
-require 'package_protections/private/multiple_namespaces_protection'
 require 'package_protections/private/visibility_protection'
 require 'package_protections/private/configuration'
